Implement uniform diff background theme config
Added a theme configuration to let the users decide whether they want
to use colorful or uniform diff backgrounds.

The colorful diff is the default using the current colors for added,
changed and deleted elements and the default background color (nord0).
The "reverse" attribute gets applied which results in a colored
background and a dark foreground.

Setting the "g:nord_uniform_diff_background" variable to "1" enables
the uniform diff background using "nord1".

GH-60