Fix wildmenu background color (menu tab completion)
The implemented feature in this GH-58 changed the background color of
the status bar ("StatusLine") group to "nord3", but the "WildMenu" group
also useed "nord3" as background color for the current tab selection.

This commit changes the "WildMenu" background color to "nord1" which is
available for both terminal ("nord1_term") and GUI ("nord1_gui") mode.

Related to GH-58

GH-64