Add configuration for italic text
The configuration to enable italic comments (1) is like the name suggests
limited to comments only, but there are other use cases like the
italic style in Markdown (2).

This commit adds a new theme configuration to explicitly enable italics.
The configurations is enabled by default when running in "gui" mode and
is disabled by default in "term" mode.

The reason for this standard behavior is the fact that there are still
shells and terminals which are not able to handle italics. Enabling
italics for terminals by default would cause unpredictable bugs and
graphical glitches.

The new configuration is named "nord_italic".

References:

  (1) https://github.com/arcticicestudio/nord-vim#italic-comments
  (2) https://github.com/arcticicestudio/nord-vim/issues/84

GH-88