Add theme config for bolder vertical split line
Previously the `VertSplit` (1) (`:help VertSplit`) key used `nord1` as
background color by default making the line appear to be very lumpy.
This commit changes this style to use `nord0` as bavkground instead to
visually merge with the background so only the separator charaters are
a visual indiciator for the split line whih makes it look more
lightweight and unclutters the overall appeareance.

To allow user who liked the previous implementation to keep the style
this commit also adds a new `nord_bold_vertical_split_line` theme config
that can be assigned to `1` to achieve the legacy design.

The README also include information and hints how to change the
separator character by customizing Vim's `fillchars` (2)
(`:help fillchars`) variable.

References:
  (1) http://vimdoc.sourceforge.net/htmldoc/syntax.html#hl-VertSplit
  (2) http://vimdoc.sourceforge.net/htmldoc/options.html#'fillchars'

GH-132