Theme configuration for bold font style rendering
Implemented a nw theme configuration for the rendering of bold font
styles.

This commit adds the new `nord_bold` theme configuration to explicitly
toggle bold font rendering styles.
It is enabled by default when running for both in GUI and terminal mode.

The reason for this standard behavior is the fact that most terminals
and shells are capable to handle bold fonts.

GH-143