Support for `php.vim` plugin
This is a follow up for arcticicestudio/nord-vim#218 [1] which added
highlighting groups for the bundled PHP syntax, but the groups are
actually defined by the StanAngeloff/php.vim [2] plugin.
Therefore the added highlighting calls will be moved to a plugin
section.
Additionally, the `phpClassExtends` and `phpClassImplements` groups
have been added to improve the highlighting for classes that implement
or extended interfaces/classes. The `phpUseClass` has also been added to
improve the highlighting for imports.

[1]: https://github.com/arcticicestudio/nord-vim/pull/218
[2]: https://github.com/StanAngeloff/php.vim

Related to GH-218
GH-262

Co-authored-by: Sven Greb <development@svengreb.de>