Remove `TSError` syntax highlighting group (#270)
The `TSError` group is used to highlight syntax/parser errors [1] which
caused an aggressive styling where the background color of many syntax
elements was rendered with `nord11` during typing. This was caused due
to the fast processing of `tree-sitter` which also resulted in highlight
flickering.
This is a known problem and was fixed by many other themes (e.g. Dracula
[2]) by removing the group again. One of the core maintainers of
`nvim-treesitter` provided a solution by remapping groups [3] and also
mentioned that the group is styled by the `nvim-treesitter` plugin but
the active theme [4].

Syntax errors can still be highlighted through linters and parsers like
Neovim's LSP [5] can still be used instead to highlight errors with the
correct style (e.g. only change the foreground color of a single word).

[1]: https://github.com/nvim-treesitter/nvim-treesitter/blob/fb5d6e04/doc/nvim-treesitter.txt#L493-L495
[2]: https://github.com/dracula/vim/pull/232
[3]: https://github.com/nvim-treesitter/nvim-treesitter/issues/78#issuecomment-647140700
[4]: https://github.com/nvim-treesitter/nvim-treesitter/issues/1016#issuecomment-797049591
[5]: https://github.com/neovim/nvim-lspconfig

Fixes GH-269