Revert conditional plugin loading
The changes introduced in PR GH-294 [1] did not take into account that
the order how plugins are loaded are not always constant and can also
change based on how users order plugins in their configurations.
When a supported plugin is loaded after Nord the global `loaded_*`
variable might not be available yet, causing the styles to be skipped
due to the conditional block guard. Also each plugin manager handles the
plugin loading order differently which is also a problem when checking
for global variable existence.

[1]: https://github.com/arcticicestudio/nord-vim/pull/294
[2]: https://github.com/arcticicestudio/nord-vim/issues/306

Fixes GH-306